refactor(web): split recurring workspace into focused modules
recurring-payments-workspace.tsx (2105 lines) becomes: shared building
blocks (recurring-payments-shared), the collection history card
(recurring-payment-collection-history), the detail workspace
(recurring-payment-detail-workspace), and the list workspace keeping
the original filename. Code moved verbatim; detail page import
repointed.
